+++ R/LearnerClustCLARA.R
@@ -6,11 +6,9 @@
 #' Clustering Large Applications (CLARA) clustering.
 #' Calls [cluster::clara()] from package \CRANpkg{cluster}.
 #'
-#' CLARA extends the PAM algorithm to handle larger datasets by working on sub-datasets of fixed size.
-#' The `k` parameter is set to 2 by default since [cluster::clara()]
-#' doesn't have a default value for the number of clusters.
-#' The predict method uses [clue::cl_predict()] to compute the
-#' cluster memberships for new data.
+#' CLARA extends the PAM algorithm to handle larger datasets by working on sub-datasets of fixed size. The `k` parameter
+#' is set to 2 by default since [cluster::clara()] doesn't have a default value for the number of clusters. The predict
+#' method uses [clue::cl_predict()] to compute the cluster memberships for new data.
 #'
 #' @templateVar id clust.clara
 #' @template learner
@@ -42,7 +40,7 @@ LearnerClustCLARA = R6Class(
         correct.d = p_lgl(default = TRUE, tags = "train")
       )
 
-      param_set$set_values(k = 2L)
+      param_set$set_values(k = 2L, keep.data = FALSE)
 
       super$initialize(
         id = "clust.clara",