+++ R/LearnerClustCMeans.R
@@ -6,10 +6,8 @@
 #' Fuzzy c-means clustering.
 #' Calls [e1071::cmeans()] from package \CRANpkg{e1071}.
 #'
-#' The `centers` parameter is set to 2 by default since [e1071::cmeans()]
-#' doesn't have a default value for the number of clusters.
-#' The predict method uses [clue::cl_predict()] to compute the
-#' cluster memberships for new data.
+#' The `centers` parameter is set to 2 by default since [e1071::cmeans()] doesn't have a default value for the number of
+#' clusters. The predict method uses [clue::cl_predict()] to compute the cluster memberships for new data.
 #'
 #' @templateVar id clust.cmeans
 #' @template learner