+++ R/LearnerClustDiana.R
@@ -6,9 +6,8 @@
 #' Divisive hierarchical clustering.
 #' Calls [cluster::diana()] from package \CRANpkg{cluster}.
 #'
-#' The predict method uses [stats::cutree()] which cuts the tree resulting from
-#' hierarchical clustering into specified number of groups (see parameter `k`).
-#' The default value for `k` is 2.
+#' The predict method uses [stats::cutree()] which cuts the tree resulting from hierarchical clustering into specified
+#' number of groups (see parameter `k`). The default value for `k` is 2.
 #'
 #' @templateVar id clust.diana
 #' @template learner
@@ -30,7 +29,7 @@ LearnerClustDiana = R6Class(
         metric = p_fct(c("euclidean", "manhattan"), default = "euclidean", tags = "train"),
         stand = p_lgl(default = FALSE, tags = "train"),
         trace.lev = p_int(0L, default = 0L, tags = "train"),
-        k = p_int(1L, default = 2L, tags = c("train", "predict"))
+        k = p_int(1L, default = 2L, tags = c("train", "cutree", "predict"))
       )
 
       param_set$set_values(k = 2L)
@@ -50,15 +49,15 @@ LearnerClustDiana = R6Class(
 
   private = list(
     .train = function(task) {
-      pv = self$param_set$get_values(tags = "train")
+      ps = self$param_set
       m = invoke(
         cluster::diana,
         x = task$data(),
         diss = FALSE,
-        .args = remove_named(pv, "k")
+        .args = remove_named(ps$get_values(tags = "train"), "k")
       )
       if (self$save_assignments) {
-        self$assignments = stats::cutree(m, pv$k)
+        self$assignments = invoke(stats::cutree, tree = m, .args = ps$get_values(tags = c("train", "cutree")))
       }
       m
     },
@@ -70,7 +69,8 @@ LearnerClustDiana = R6Class(
       }
 
       warn_prediction_useless(self$id)
-      partition = self$assignments %??% stats::cutree(self$model, pv$k)
+      partition = self$assignments %??%
+        invoke(stats::cutree, tree = self$model, .args = self$param_set$get_values(tags = c("train", "cutree")))
 
       PredictionClust$new(task = task, partition = partition)
     }