+++ R/LearnerClustFeatureless.R
@@ -4,8 +4,7 @@
 #'
 #' @description
 #' Featureless clustering.
-#' Randomly (but evenly) assigns observations to `num_clusters` partitions
-#' (default: 1 partition).
+#' Randomly (but evenly) assigns observations to `num_clusters` partitions (default: 1 partition).
 #'
 #' @templateVar id clust.featureless
 #' @template learner